+++ www/scripts/opennms/services/Modals.js
@@ -66,9 +66,7 @@
 					modal.scope.hasAddress = true;
 				}
 
-				NodeService.canSetLocation().then(function(isValid) {
-					modal.scope.canUpdateGeolocation = isValid;
-				});
+				modal.scope.canUpdateGeolocation = Info.canSetLocation();
 
 				var avail = AvailabilityService.node(modal.scope.node.id).then(function(results) {
 					//console.log('AvailabilityService got results:',results);
@@ -180,7 +178,7 @@
 						break;
 				}
 			});
-			modal.scope.$on('opennms.settings.changed', function(ev, newSettings, oldSettings, changedSettings) {
+			modal.scope.$on('opennms.settings.updated', function(ev, newSettings, oldSettings, changedSettings) {
 				if (timer && changedSettings && changedSettings.refreshInterval) {
 					modal.scope.updateData();
 				}
@@ -254,7 +252,8 @@
 			};
 
 			modal.scope.refreshOutages = function() {
-				OutageService.summaries().then(function(outages) {
+				var promise = OutageService.summaries();
+				promise.then(function(outages) {
 					var oldOutages = byId(modal.scope.outages);
 					var oldOutage;
 					for (var i=0; i < outages.length; i++) {
@@ -265,8 +264,14 @@
 						getOutages(outages[i]);
 					}
 					modal.scope.outages = outages;
+					delete modal.scope.error;
+					modal.scope.$broadcast('scroll.refreshComplete');
+				}, function(err) {
+					modal.scope.error = err;
+					modal.scope.outages = [];
 					modal.scope.$broadcast('scroll.refreshComplete');
 				});
+				return promise;
 			};
 
 			var startRefresh = function() {
@@ -282,13 +287,14 @@
 
 			modal.scope.show = function() {
 				modal.scope.outages = [];
-				startRefresh();
-				modal.show();
+				modal.scope.refreshOutages().then(function() {
+					modal.show();
+				});
 			};
 			modal.scope.hide = function() {
 				modal.hide();
 			};
-			modal.scope.$on('opennms.settings.changed', function(ev, newSettings, oldSettings, changedSettings) {
+			modal.scope.$on('opennms.settings.updated', function(ev, newSettings, oldSettings, changedSettings) {
 				if (timer && changedSettings && changedSettings.refreshInterval) {
 					stopRefresh();
 					startRefresh();
@@ -394,7 +400,7 @@
 			modal.scope.hide = function() {
 				modal.hide();
 			};
-			modal.scope.$on('opennms.settings.changed', function(ev, newSettings, oldSettings, changedSettings) {
+			modal.scope.$on('opennms.settings.updated', function(ev, newSettings, oldSettings, changedSettings) {
 				if (timer && changedSettings && changedSettings.refreshInterval) {
 					stopRefresh();
 					startRefresh();
@@ -451,12 +457,8 @@
 			};
 			modal.scope.show = function() {
 				modal.scope.errors = Errors.get();
-				Info.get().then(function(info) {
-					modal.scope.info = info;
-				});
-				NodeService.canSetLocation().then(function(canSet) {
-					modal.scope.canSetLocation = canSet;
-				});
+				modal.scope.info = Info.get();
+				modal.scope.canSetLocation = Info.canSetLocation();
 				AvailabilityService.supported().then(function(isSupported) {
 					modal.scope.hasAvailability = isSupported;
 				});